refactor: moved query statements to own file #64
