+++ src/parser/metrics/Functions.ts
@@ -1,15 +1,12 @@
 import { QueryBuilder } from "../queries/QueryBuilder";
-import {
-    ExpressionMetricMapping,
-    QueryStatementInterface,
-    SimpleQueryStatement,
-} from "../helper/Model";
+import { ExpressionMetricMapping } from "../helper/Model";
 import { getQueryStatements } from "../helper/Helper";
 import { FileMetric, Metric, MetricResult, ParseFile } from "./Metric";
 import { debuglog, DebugLoggerFunction } from "node:util";
 import { QueryMatch } from "tree-sitter";
 import Parser from "tree-sitter";
 import { Languages } from "../helper/Languages";
+import { QueryStatementInterface, SimpleQueryStatement } from "../helper/QueryStatements";
 
 let dlog: DebugLoggerFunction = debuglog("metric-gardener", (logger) => {
     dlog = logger;
