+++ src/parser/resolver/fullyQualifiedTypeNames/resolverStrategy/QueryStrategy.ts
@@ -3,9 +3,9 @@ import { TreeParser } from "../../../helper/TreeParser";
 import { QueryBuilder } from "../../../queries/QueryBuilder";
 import { formatCaptures } from "../../../helper/Helper";
 import { ParseFile } from "../../../metrics/Metric";
-import { SimpleQueryStatement } from "../../../helper/Model";
 import { debuglog, DebugLoggerFunction } from "node:util";
 import { QueryCapture } from "tree-sitter";
+import { SimpleQueryStatement } from "../../../helper/QueryStatements";
 
 let dlog: DebugLoggerFunction = debuglog("metric-gardener", (logger) => {
     dlog = logger;
